metodos en model Proyecto cambiados

+++ proyecto/views.py
@@ -24,7 +24,7 @@ class listarProyectos(LoginRequiredMixin, PermissionListMixin, ListView):
        **Listar Proyecto:**
         03/09/2021
         Vista utilizada para crear listar los proyectos
-        creados en el sistema, el usuario debe tener el permiso de ver el proyecto.
+        creados en el sistema, el usuario debe tener permiso en cada proyecto para que estos sean visibles a el.
     """
     model = Proyecto
     permission_required = "VER_PROYECTO"
@@ -54,40 +54,6 @@ def editarProyecto(request, proyecto_id):
     # print("editar ->",context)
     return render(request, "proyecto/editar.html", context)
 
-def setScrum(request, proyecto, user, rol):
-    """
-       **Añadir Srum Master:**
-        03/09/2021
-        Funcion utilizada para añadir el scrum Master al proyecto .
-        Solicita el proyecto, user y el rol para añadir al proyecto
-
-    """
-    if (request.method == 'POST'):
-        formMiembro = setMiembroForms(request.POST)
-        formMiembro.instance.proyectos = proyecto
-        formMiembro.instance.miembro = user
-        formMiembro.instance.rol = rol
-        formMiembro.instance.save()
-        user.groups.add(rol.group)
-    else:
-        formMiembro = setMiembroForms()
-
-def scrumRol(proyecto_id):
-    """
-       **Añadir el rol al Scrum Master:**
-        03/09/2021
-        Funcion utilizada para añadir el rol al Scrum Master .
-        Solicita el id del proyecto, para luego crear el rol mediante
-        el concatenacion del id del proyecto y luego añadirle al mismo
-    """
-    grupo = Group.objects.create(name='Scrum' + str(proyecto_id))
-    proyecto_actual = Proyecto.objects.get(pk=proyecto_id)
-    rol = Rol.objects.create_rol('Scrum' + str(proyecto_id), grupo)
-    rol.save()
-    for perm in permissions:
-        assign_perm(perm[0], rol.group, proyecto_actual)
-    proyecto_actual.roles.add(rol)
-    return (rol)
 
 @login_required(login_url='/login')
 def crearProyecto(request):
@@ -98,30 +64,22 @@ def crearProyecto(request):
         Requiere que el usuario este logeado
 
     """
-    proyecto = Proyecto.objects.all()
-    miembro = Miembro.objects.all()
-    # context = {}
     if request.method == "POST":
         form = ProyectoCrearForms(request.POST)
         if form.is_valid():
             data = form.cleaned_data
             nombre_proyecto = data['nombre_proyecto']
-            proyecto = Proyecto.objects.filter(nombre_proyecto =nombre_proyecto)
-            if not proyecto.exists():
-                data = form.cleaned_data
-                nombre = data['nombre_proyecto']
+            proyecto_actual = Proyecto.objects.filter(nombre_proyecto =nombre_proyecto)
+            if not proyecto_actual.exists():
                 user = data['scrum_master']
                 form.save()
-                proyecto = Proyecto.objects.get(nombre_proyecto=nombre)
-                rol = scrumRol(proyecto.pk)
-                setScrum(request, proyecto, user, rol)
+                proyecto_actual = Proyecto.objects.get(nombre_proyecto=nombre_proyecto)
+                rol = Rol.crearScrum(proyecto_actual.pk)
+                proyecto_actual.setScrum(user, rol)
                 return redirect("listarProyectos")
-
-                context = {'proyectos': proyecto, 'miembros': miembro}
-                return render(request, "listarProyectos", context)
             else:
                 error = True
-                context = {'error': error, 'form': form, 'proyectos': proyecto, 'miembros': miembro}
+                context = {'error': error, 'form': form, 'proyectos': proyecto}
                 return render(request, "proyecto/crearProyecto.html", context)
 
     else:
@@ -137,7 +95,6 @@ def proyecto(request, proyecto_id):
         Vista utilizada recibir a los usuarios en un proyecto.
         Solicita el id del proyecto
     """
-
     proyecto = Proyecto.objects.get(pk=proyecto_id)
     try:
         miembro= Miembro.objects.get(proyectos=proyecto_id, miembro=request.user)
@@ -145,8 +102,6 @@ def proyecto(request, proyecto_id):
         es_scrum=False
         if miembro.rol.nombre == scrum:
             es_scrum=True
-
-        # print(proyecto.nombre_proyecto,proyecto.pk)
         context = {'proyecto_id': proyecto_id,
                    'proyecto': proyecto,'es_scrum':es_scrum}
 
@@ -215,34 +170,22 @@ def crearGrupo(request, proyecto_id):
     if request.method == "POST":
         form = CrearGrupo(request.POST or None)
         if form.is_valid():
-
             data = form.cleaned_data
             nombre = data['nombre']
             permisos_elegidos = data['permisos_proyecto']
 
             if Rol.objects.filter(nombre=nombre).exists():
-                # No se puede crear el rol ya que existe uno con ese nombre
                 print('Un rol ya existe con ese nombre')
             else:
                 # No existe un grupo con el mismo
-
-                grupo = Group.objects.create(name=nombre)
                 proyecto_actual = Proyecto.objects.get(pk=proyecto_id)
-
-                rol = Rol.objects.create_rol(nombre, grupo)
-                proyecto_actual.roles.add(rol)
-
-                for perm in permisos_elegidos:
-                    assign_perm(perm, rol.group, proyecto_actual)
-
-                rol.save()
+                Rol.crear(nombre, permisos_elegidos, proyecto_actual)
             return redirect(reverse('listaRol', kwargs={'proyecto_id': proyecto_id}))
     else:
         form = CrearGrupo()
-    # grupos=Group.objects.all()
     proyecto = Proyecto.objects.get(pk = proyecto_id)
-    permisos = Permission.objects.all()
-    context = {'Permisos': permisos, 'form': form, 'proyecto_id': proyecto_id , 'proyecto': proyecto}
+
+    context = {'form': form, 'proyecto_id': proyecto_id , 'proyecto': proyecto}
     return render(request, "rol/crear.html", context)
 
 
@@ -290,53 +233,17 @@ def editarRol(request, rol_id, proyecto_id):
 
 
             rol_actual = Rol.objects.get(pk=rol_id)
+            rol_actual.editar(nombre, permisos_elegidos, proyecto_id)
 
-
-            # Se trae el modelo del rol
-
-            rol_actual.nombre=nombre
-            # se extraen los miembros del proyecto con el rol
-            miembros_con_el_rol = Miembro.objects.filter(rol=rol_actual)
-
-            # Se estrae el grupo del rol y se elimina
-            grupo_anterior = rol_actual.group
-            grupo_anterior.delete()
-
-            # se crea nuevo grupo
-            rol_actual.group = Group.objects.create(name=nombre)
-            #SE NECESITA EL PROYECTO_ID PARA ASIGNAR LOS NUEVOS PERMISOS
-            proyecto_actual = Proyecto.objects.get(pk=proyecto_id)
-
-            # se asignan los nuevos permisos al grupo
-            for perm in permisos_elegidos:
-                assign_perm(perm, rol_actual.group, proyecto_actual)
-            # se agregan a los miembros al nuevo grupo si es que alguno existe
-            if miembros_con_el_rol.exists():
-                for miembro in miembros_con_el_rol:
-                     miembro.miembro.groups.add(rol_actual.group)
-
-
-            rol_actual.group.save()
-            rol_actual.save()
-            proyecto_actual.roles.add(rol_actual)
-            proyecto_actual.save()
             return redirect(reverse('listaRol', kwargs={'proyecto_id': proyecto_id}))
-            # proyecto_actual = Proyecto.objects.get(pk=proyecto_id)
-            #
-            # rol = Rol.objects.create_rol(nombre, grupo)
-            # proyecto_actual.roles.add(rol)
-            #
-            # rol.save()
-            # for perm in permisos_elegidos:
-            #     assign_perm(perm, rol.group, proyecto_actual)
 
     else:
         form = EditarGrupo()
         proyecto = Proyecto.objects.get(id=proyecto_id)
         rol_nombre = Rol.objects.get(id=rol_id)
-    context = {"rol_id": rol_id, "proyecto_id":proyecto_id ,"form": form, 'proyecto':proyecto, 'rol_nombre':rol_nombre}
+        context = {"rol_id": rol_id, "proyecto_id":proyecto_id ,"form": form, 'proyecto':proyecto, 'rol_nombre':rol_nombre}
 
-    return render(request, "rol/editar.html", context)
+        return render(request, "rol/editar.html", context)
 
 
 def eliminarRol(request, rol_id, proyecto_id):
@@ -401,8 +308,6 @@ def editar_rolmiembro(request, proyecto_id, miembro_id):
     return render(request, "proyecto/miembroEditar.html", context)
 
 
-
-
 def eliminarProyecto(request,proyecto_id):
     """
        **Eliminar Proyecto:**
@@ -412,9 +317,11 @@ def eliminarProyecto(request,proyecto_id):
     """
     print(proyecto_id)
     proyecto=Proyecto.objects.get(id=proyecto_id)
-    proyecto.delete()
-
+    if proyecto.estado=='PENDIENTE':
+        proyecto.delete()
+        Miembro.objects.filter(proyecto_id=proyecto_id).delete()
     return redirect("/home/proyectos/")
+
 
 def iniciarProyecto(request, proyecto_id):
     """
@@ -424,7 +331,6 @@ def iniciarProyecto(request, proyecto_id):
         Solicita el id del proyecto
     """
     proyecto=Proyecto.objects.get(pk=proyecto_id)
-    miembros=Miembro.objects.filter(proyectos=proyecto)
-    proyecto.estado="ACTIVO"
-    context = {"proyecto_id": proyecto_id,"proyecto": proyecto, "miembros": miembros}
+    proyecto.iniciar_proyecto()
+    context = {"proyecto_id": proyecto_id,"proyecto": proyecto}
     return render(request,"desarrollo/desarrollo.html", context)
